Users are no longer allowed to delete the last visit from an arm on the calendar. [#75626804]

+++ app/assets/javascripts/portal/fulfillment2.js.coffee
@@ -325,29 +325,32 @@ $(document).ready ->
         $("#submit_visit").attr("disabled", false).removeClass("ui-state-disabled")
 
   $(document).on('click', '.delete_visit_link', ->
-    sr_id = $(this).data('service_request_id')
-    data =
-      'sub_service_request_id': $(this).data('sub_service_request_id')
-      'service_request_id': sr_id
-      'visit_position': $('#delete_visit_position').val()
-      'arm_id': $('#arm_id').val()
-      'study_tracker': $('#study_tracker_hidden_field').val() || null
-    $.ajax
-      type: 'DELETE'
-      url:   "/portal/admin/service_requests/#{sr_id}/remove_per_patient_per_visit_visit"
-      data:  JSON.stringify(data)
-      dataType: 'script'
-      contentType: 'application/json; charset=utf-8'
-      success: ->
-        $().toastmessage('showSuccessToast', "Service request has been saved.")
+    if $(this).data('visit_count') <= 1
+      alert("You can't delete the last visit. Please delete the arm if visits are no longer required.")
+    else
+      sr_id = $(this).data('service_request_id')
+      data =
+        'sub_service_request_id': $(this).data('sub_service_request_id')
+        'service_request_id': sr_id
+        'visit_position': $('#delete_visit_position').val()
+        'arm_id': $('#arm_id').val()
+        'study_tracker': $('#study_tracker_hidden_field').val() || null
+      $.ajax
+        type: 'DELETE'
+        url:   "/portal/admin/service_requests/#{sr_id}/remove_per_patient_per_visit_visit"
+        data:  JSON.stringify(data)
+        dataType: 'script'
+        contentType: 'application/json; charset=utf-8'
+        success: ->
+          $().toastmessage('showSuccessToast', "Service request has been saved.")
 
-      error: (jqXHR, textStatus, errorThrown) ->
-        if jqXHR.status == 500 and jqXHR.getResponseHeader('Content-Type').split(';')[0] == 'text/javascript'
-          errors = JSON.parse(jqXHR.responseText)
-        else
-          errors = [textStatus]
-        for error in errors
-          $().toastmessage('showErrorToast', "#{error.humanize()}.");
+        error: (jqXHR, textStatus, errorThrown) ->
+          if jqXHR.status == 500 and jqXHR.getResponseHeader('Content-Type').split(';')[0] == 'text/javascript'
+            errors = JSON.parse(jqXHR.responseText)
+          else
+            errors = [textStatus]
+          for error in errors
+            $().toastmessage('showErrorToast', "#{error.humanize()}.");
   )
 
   $(document).on('click', '#add_service', ->
